+++ ownership-chain/pallets/parachain-staking/src/lib.rs
@@ -2197,82 +2197,4 @@ pub mod pallet {
 			Self::selected_candidates().into_inner()
 		}
 	}
-
-	// Modification by Freeverse
-
-	use frame_support::traits::EstimateNextSessionRotation;
-	use sp_runtime::Permill;
-	use sp_staking::SessionIndex;
-
-	type AccountIdOf<T> = <T as frame_system::Config>::AccountId;
-
-	// impl<T: Config> pallet_session::SessionManager<AccountIdOf<T>> for Pallet<T> {
-	// 	/// 1. A new session starts.
-	// 	/// 2. In hook new_session: Read the current top n candidates from the TopCandidates and
-	// 	///    assign this set to author blocks for the next session.
-	// 	/// 3. AuRa queries the authorities from the session pallet for this session and picks
-	// 	///    authors on round-robin-basis from list of authorities.
-	// 	fn new_session(new_index: SessionIndex) -> Option<Vec<AccountIdOf<T>>> {
-	// 		log::warn!(
-	// 			"assembling new collators for new session {} at #{:?}",
-	// 			new_index,
-	// 			<frame_system::Pallet<T>>::block_number(),
-	// 		);
-
-	// 		let collators = Pallet::<T>::selected_candidates().to_vec();
-	// 		if collators.is_empty() {
-	// 			// we never want to pass an empty set of collators. This would brick the chain.
-	// 			log::error!("💥 keeping old session because of empty collator set!");
-	// 			None
-	// 		} else {
-	// 			Some(collators)
-	// 		}
-	// 	}
-
-	// 	fn end_session(_end_index: SessionIndex) {
-	// 		// we too are not caring.
-	// 	}
-
-	// 	fn start_session(_start_index: SessionIndex) {
-	// 		// we too are not caring.
-	// 	}
-	// }
-
-	impl<T: Config> pallet_session::ShouldEndSession<u32> for Pallet<T> {
-		fn should_end_session(now: u32) -> bool {
-			let round = <Round<T>>::get();
-			// always update when a new round should start
-			round.should_update(now as u64)
-		}
-	}
-
-	impl<T: Config> EstimateNextSessionRotation<u32> for Pallet<T> {
-		fn average_session_length() -> u32 {
-			<Round<T>>::get().length
-		}
-
-		fn estimate_current_session_progress(now: u32) -> (Option<Permill>, Weight) {
-			let round = <Round<T>>::get();
-			// TODO check the try_into
-			let passed_blocks = now.saturating_sub(round.first.try_into().unwrap());
-
-			(
-				Some(Permill::from_rational(passed_blocks, round.length)),
-				// One read for the round info, blocknumber is read free
-				T::DbWeight::get().reads(1),
-			)
-		}
-
-		fn estimate_next_session_rotation(_now: u32) -> (Option<u32>, Weight) {
-			let round = <Round<T>>::get();
-
-			// TODO check this try_into
-			let first_round: u32 = round.first.try_into().unwrap();
-			(
-				Some(first_round + round.length),
-				// One read for the round info, blocknumber is read free
-				T::DbWeight::get().reads(1),
-			)
-		}
-	}
 }
