+++ ownership-chain/runtime/laos/src/lib.rs
@@ -401,8 +401,8 @@ impl pallet_session::Config for Runtime {
 	type ValidatorId = <Self as frame_system::Config>::AccountId;
 	// we don't have stash and controller, thus we don't need the convert as well.
 	type ValidatorIdOf = ConvertInto;
-	type ShouldEndSession = ParachainStaking;
-	type NextSessionRotation = ParachainStaking;
+	type ShouldEndSession = configs::parachain_staking::ShouldEndSession;
+	type NextSessionRotation = configs::parachain_staking::NextSessionRotation;
 	type SessionManager = configs::parachain_staking::SessionManager;
 	// Essentially just Aura, but let's be pedantic.
 	type SessionHandler = <SessionKeys as sp_runtime::traits::OpaqueKeys>::KeyTypeIdProviders;
